refactor(gateway): split providers.rs into per-provider submodules
Decompose the 2,067-line gateway/providers.rs into a directory module:

- providers/mod.rs      (868) — shared helpers: error/retry, SSE senders,
                                 resolve_request, compact_conversation,
                                 validate_model_connection, history formatting
- providers/openai.rs   (789) — Chat Completions + Responses API, SSE consumers
- providers/anthropic.rs (339) — Messages API + parse_anthropic_response
- providers/google.rs   (113) — Gemini

Public API is preserved via re-exports, so all `providers::call_*` call sites
in gateway/mod.rs resolve identically — zero call-site changes. Only
find_event_boundary widened to pub(crate) for the SSE consumers.

cargo clippy --workspace --all-targets -- -D warnings: clean
cargo fmt --all -- --check: clean
